feat(scpn-fusion): data standardization & interoperability — 6-phase upgrade
Phase 5: mypy strict expansion (6→17 files), PEP 561 py.typed marker
Phase 1: IMAS Data Dictionary v3 compliant IDS converters (equilibrium,
         core_profiles, summary), MDSplus mock tests, write_ids/read_ids IO
Phase 2: ECE radiometer, soft X-ray brightness, bolometry forward models
Phase 4: GMRES + multigrid Criterion benchmarks, Python solver comparison
Phase 3: Boris particle integrator PyO3 bridge (PyParticle, advance, heating)
Phase 6: README solver performance table with head-to-head comparisons

1067 tests pass, 0 mypy errors across 17 strict-mode files, Rust compiles clean.

+++ src/scpn_fusion/core/stability_mhd.py
@@ -175,7 +175,7 @@ def mercier_stability(qp: QProfile) -> MercierResult:
             break
 
     return MercierResult(
-        rho=qp.rho, D_M=D_M, stable=stable,
+        rho=qp.rho, D_M=D_M.astype(np.float64), stable=stable,
         first_unstable_rho=first_unstable_rho,
     )
